squash: break out formFields.js
